refactor: move computeInterfaceId.ts
