+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./computeTitleEscrowAddress";
 export * from "./encodeInitParams";
 export * from "./isSupportedTitleEscrowFactory";
 export * from "./getEventFromTransaction";
+export * from "./computeInterfaceId";
